backend/requirements.txt
fastapi
pydantic
passlib[bcrypt]
python-jose
uvicorn
pyjwt